Changed file composition and added contact
Made a wish list page with the content that was
in app component, I want the app to be free.
Made the contact component and with router, routed
the wish list page and contact page to each of
their urls. Wish to main and contact to /contact.
Contact is to put name, email and a message. Also
implemented some errors to be shown.